refactor: move method search into search-api service (DD-7)
