+++ src/app/features/home/interfaces/city-coordinates.interface.ts
@@ -1,0 +1,4 @@
+export interface CityCoordinates {
+  latitude: number;
+  longitude: number;
+}
