+++ src/app/features/home/interfaces/city-info.interface.ts
