+++ src/app/features/home/services/location-api/location-api.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { inject, Injectable } from '@angular/core';
 import { environment } from 'src/environment';
-import { CityInfo } from '../../interfaces/city-info';
+import { CityInfo } from '../../interfaces/city-info.interface';
 
 const responseLimit = 5;
